requirements.txt
python-telegram-bot==22.7
python-dotenv==1.0.0
fastapi==0.111.0
uvicorn==0.29.0
pandas==2.2.2
openpyxl==3.1.2
psycopg2-binary==2.9.9
